Added Line dection code.
I've modified Yonnas' touch.py to work with my homography code. All we need to is accept a frame of the video as the destination image (dst_path), the src path is just the birdseye view of the field.

One we have manual tracking done, I can integrate the automated tracking.